Updated error explanation by adding the output from cargo tarpaulin to 'build_output.txt'
There is at least one person trying to reproduce the error that I'm
getting, but they aren't having much success doing so.  This commit
provides the output of cargo tarpaulin with full backtrace, and at a log
level of 'trace', which I'm hoping will yield enough information to
determine what the problem is.